refactor: rename files to follow kebab-case and update imports
