+++ lib/entry-list.ts
