+++ lib/types/entry-list.types.ts
